pyproject.toml
[project]
    authors = [
        { name = "Yifei Liu", email = "34564223+EzekielDaun@users.noreply.github.com" },
    ]
    dependencies = [
        "jaxlie (>=1.5.0,<2.0.0)",
        "jaxtyping (>=0.3.2,<0.4.0)",
        "networkx (>=3.5,<4.0)",
    ]
    description = ""
    name = "general-manipulator-kinematics"
    readme = "README.md"
    requires-python = ">=3.12,<3.14"
    version = "0.0.3"

[project.optional-dependencies]
    matplotlib-viz = ["matplotlib (>=3.10.5,<4.0.0)"]
    pyqtgraph-viz = [
        "pyopengl (>=3.1.9,<4.0.0)",
        "pyqtgraph (>=0.13.7,<0.14.0)",
        "pyside6 (<7.0.0)",
    ]

[build-system]
    build-backend = "hatchling.build"
    requires      = ["hatchling"]

[tool.hatch.build.targets.wheel]
    packages = ["src/general_manipulator_kinematics"]
